+++ src/pat/auto-suggest/tests.js
@@ -154,6 +154,59 @@ define(["pat-autosuggest"], function(pattern) {
                 utils.removeSelect2();
             });
         });
+
+
+        describe("Placeholder tests", function () {
+            beforeEach(function() {
+                $("div#lab").remove(); // Looks likes some specs don't clean up after themselves.
+                $("<div/>", {id: "lab"}).appendTo(document.body);
+            });
+
+            afterEach(function() {
+                $("#lab").remove();
+            });
+
+            it("A placeholder on the original element is reused on the auto-suggest element", function() {
+
+                var placeholder = "Test placeholder";
+
+                $("<input/>", {
+                    "id": "select2",
+                    "class": "pat-autosuggest",
+                    "placeholder": "Test placeholder",
+                }).appendTo($("div#lab"));
+
+                var $el = $("input.pat-autosuggest");
+
+                expect($el.prop('placeholder')).toBe(placeholder);
+                pattern.init($el);
+                // XXX Placeholder is used as value of the input field.
+                // Change this as soon Select2 changes this odd behavior.
+                expect($(".select2-input").val()).toBe(placeholder);
+                utils.removeSelect2();
+            });
+
+            it("No placeholder doesn't create an automatic one.", function() {
+
+                $("<input/>", {
+                    "id": "select2",
+                    "class": "pat-autosuggest",
+                }).appendTo($("div#lab"));
+
+                var $el = $("input.pat-autosuggest");
+
+                expect($el.prop('placeholder')).toBe('');
+                pattern.init($el);
+                // XXX Placeholder is used as value of the input field.
+                // Change this as soon Select2 changes this odd behavior.
+                expect($(".select2-input").val()).toBe('');
+                utils.removeSelect2();
+            });
+
+
+        });
+
+
     });
 
 });